backend/pyproject.toml
[project]
name = "maayu"
version = "0.1.0"
description = "A suite of MCP servers and extraction tools for Maayu AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.37.29",
    "fastapi[standard]>=0.112.1",
    "google>=3.0.0",
    "google-genai>=1.9.0",
    "langchain-google-genai>=2.1.2",
    "langchain-mcp-adapters>=0.0.7",
    "langgraph>=0.3.25",
    "openai>=1.71.0",
    "pydantic-settings>=2.8.1",
    "pymupdf>=1.25.5",
    "qdrant-client>=1.13.3",
    "rich>=13.7.1",
    "ruff>=0.11.4",
    "structlog>=25.2.0",
]
